capture stderr from tf init and send it back to the controller
